% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yugioh_database_version.R
\name{get_yugioh_database_version}
\alias{get_yugioh_database_version}
\title{Get Yugioh Database Version}
\usage{
get_yugioh_database_version()
}
\value{
A list containing the database version and last update information.
}
\description{
This function retrieves the current version of the Yugioh database.
}
\examples{
get_yugioh_database_version()
}
